requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.3
langchain-google-genai==0.0.11
langchain-core==0.1.33
